backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
fastapi-mail==1.4.1
jinja2==3.1.2
aiohttp==3.9.1
pydantic-settings==2.1.0